Added: First completed version of i18n connector for ubiquo_media (and required changes to existing files)

+++ lib/ubiquo_media/connectors/i18n.rb
@@ -1,19 +1,26 @@
 module UbiquoMedia
   module Connectors
-    class Standard < Base
+    class I18n < Base
       
       
       module Asset
         
         def self.included(klass)
           klass.send(:extend, ClassMethods)
-          Standard.register_uhooks klass, ClassMethods
+          klass.send(:translatable, :name)
+          I18n.register_uhooks klass, ClassMethods
         end
         
         module ClassMethods
+          
           # Applies any required extra scope to the filtered_search method
           def uhook_filtered_search filters = {}
-            yield
+            filter_locale = filters[:locale] ?
+              {:find => {:conditions => ["assets.locale <= ?", filters[:locale]]}} : {}
+              
+            with_scope(filter_locale) do
+              yield
+            end
           end
         end
         
@@ -39,18 +46,28 @@ module UbiquoMedia
         def self.included(klass)
           klass.send(:include, InstanceMethods)
           klass.send(:helper, Helper)
-          Standard.register_uhooks klass, InstanceMethods
+          I18n.register_uhooks klass, InstanceMethods
         end
         
         module Helper
           # Returns a string with extra filters for assets
           def uhook_asset_filters url_for_options
-            ''
+            render_filter(:links, url_for_options,
+              :caption => ::Asset.human_attribute_name("locale"),
+              :field => :filter_locale,
+              :collection => Locale.active,
+              :id_field => :iso_code,
+              :name_field => :native_name
+            )
           end
 
           # Returns an array with any display information for extra assets filters
           def uhook_asset_filters_info
-            []
+            [filter_info(
+              :string, params,
+              :field => :filter_locale,
+              :caption => ::Asset.human_attribute_name("locale"))
+            ]
           end
         end
         
@@ -58,17 +75,19 @@ module UbiquoMedia
           
           # Returns a hash with extra filters to apply
           def uhook_index_filters
-            {}
+            {:locale => params[:filter_locale]}
           end
           
           # Initializes a new instance of asset.
           def uhook_new_asset
-            ::AssetPublic.new
+            ::AssetPublic.translate(params[:from], current_locale, :copy_all => true)
           end
           
           # Creates a new instance of asset.
           def uhook_create_asset visibility
-            visibility.new(params[:asset])
+            asset = visibility.new(params[:asset])
+            asset.locale = current_locale
+            asset
           end
          
 #          #updates an asset instance. returns a boolean that means if update was done.
@@ -78,7 +97,13 @@ module UbiquoMedia
 #
           #destroys an asset instance. returns a boolean that means if the destroy was done.
           def uhook_destroy_asset(asset)
-            asset.destroy
+            destroyed = false
+            if params[:destroy_content]
+              destroyed = asset.destroy_content
+            else
+              destroyed = asset.destroy
+            end
+            destroyed
           end
         end
       end
@@ -87,12 +112,12 @@ module UbiquoMedia
         
         def self.included(klass)
           klass.send(:extend, ClassMethods)
-          Standard.register_uhooks klass, ClassMethods
+          I18n.register_uhooks klass, ClassMethods
         end
         
         module ClassMethods
           def uhook_create_assets_table
-            create_table :assets do |t|
+            create_table :assets, :translatable => true do |t|
               yield t
             end
           end
